mysite/requirements.txt
flake8==3.7.9
nox==2019.11.9
pytest==3.10.1
requests==2.22.0
Flask-Login==0.5.0
numpy==1.19.2
PyJWT==1.7.1
cryptography
gunicorn
uwsgi
markupsafe
python-crontab